Service_1/requirements.txt
Flask
Flask-SQLAlchemy==3.0.2
PyJWT==2.3.0
Werkzeug
requests
python-dotenv
flask_cors
jsonify